Optimize the way we send attribute messages to the local stats.
We now bunch up all the attributes one local stats processor is responsible
for in one message and send it on, instead of creating one message for
every single attribute.
The distribution of messages still relies on the engine's shuffle
mechanism, i.e. not 100% that there is 1:1 correspondence between
local stats PIs and messages sent, but it should be Good Enough TM.